+++ packages/d-render/src/cip-table/column-input.tsx
@@ -1,10 +1,10 @@
 import { computed, defineComponent, type PropType } from 'vue'
 import CipFormItem from '../cip-form-item'
-import type { IAnyObject, IRenderConfig } from '@d-render/shared'
+import type { IAnyObject, ITableColumnConfig } from '@d-render/shared'
 export default defineComponent({
   name: 'ColumnInput',
   props: {
-    config: { type: Object as PropType<IRenderConfig>, required: true },
+    config: { type: Object as PropType<ITableColumnConfig['config']>, required: true },
     fieldKey: { type: String },
     columnKey: String,
     index: { type: Number, required: true },
@@ -16,17 +16,20 @@ export default defineComponent({
     updateData: {
       type: Function as PropType<(val: IAnyObject, index: number)=>void>,
       required: true
-    }
+    },
+    rowEdit: { type: Boolean, default: undefined }
   },
   setup (props) {
     const computedConfig = computed(() => {
       const config = { ...props.config }
       config.width = '100%'
       config.hideLabel = true
-      if (config.writable === true) {
+      const writable = props.config.writable
+      if (writable) {
         config.ruleKey = `${props.tableRuleKey}.${props.propertyKey}.${props.columnKey}`
       }
-      if (!config.writable) {
+      if (!writable) {
+        config.writable = false
         config.readable = true
         if (!config.dynamic) {
           config.dependOn = []
@@ -43,6 +46,7 @@ export default defineComponent({
         tableData, // table的数据
         config: computedConfig.value, // 改变后的配置
         inTable: true, // 特殊标记
+        rowEdit: props.rowEdit, // 行编辑(因为需要控制configConfig导致的编辑状态变化,所以只能交给DrFormItem处理)
         'onUpdate:model': (val: IAnyObject) => { // 更新数据
           model = val
           updateData(val, index)
@@ -50,6 +54,6 @@ export default defineComponent({
       }
     })
 
-    return () => <CipFormItem {...fromItemProps.value}/>
+    return () => <CipFormItem style={'height: 30px;'} {...fromItemProps.value}/>
   }
 })