+++ play/src/views/develop/routes.js
@@ -1,0 +1,7 @@
+export const routes = [
+  {
+    name: 'DrTableBase',
+    path: '/dr-table/base',
+    component: () => import('./dr-table/base')
+  }
+]